Reuse /charge-elements as base for /jobs
`/charge-elements` only had one endpoint that was there to kick off the time-limited job. So, it made sense to rename/move it and its components to be the base for the new `/jobs` endpoint.
