+++ app/plugins/router.plugin.js
@@ -16,11 +16,11 @@ const BillLicences = require('../routes/bill-licences.routes.js')
 const BillRoutes = require('../routes/bills.routes.js')
 const BillRunRoutes = require('../routes/bill-runs.routes.js')
 const BillingAccountRoutes = require('../routes/billing-accounts.routes.js')
-const ChargeElements = require('../routes/charge-elements.routes.js')
 const CheckRoutes = require('../routes/check.routes.js')
 const DataRoutes = require('../routes/data.routes.js')
 const FilterRoutesService = require('../services/plugins/filter-routes.service.js')
 const HealthRoutes = require('../routes/health.routes.js')
+const JobRoutes = require('../routes/jobs.routes.js')
 const LicenceRoutes = require('../routes/licence.routes.js')
 const RootRoutes = require('../routes/root.routes.js')
 
@@ -34,8 +34,8 @@ const routes = [
   ...BillRoutes,
   ...BillRunRoutes,
   ...BillingAccountRoutes,
-  ...ChargeElements,
   ...LicenceRoutes,
+  ...JobRoutes,
   ...CheckRoutes,
   ...DataRoutes
 ]
