+++ app/services/jobs/time-limited/process-time-limited-licences.service.js
@@ -6,8 +6,8 @@
  */
 
 const FetchTimeLimitedLicencesService = require('./fetch-time-limited-licences.service.js')
-const { timestampForPostgres } = require('../../lib/general.lib.js')
-const Workflow = require('../../models/workflow.model.js')
+const { timestampForPostgres } = require('../../../lib/general.lib.js')
+const Workflow = require('../../../models/workflow.model.js')
 
 /**
  * Puts SROC licences into workflow that have a related `purpose` that is due to expire in less than 50 days
